Added tests, reworked for sanity and executing 1 frame at a time
This is the running online version now (except that it is missing the MSG fixes not updated yet)
A rework was done to be able to run 1 action every frame. Making health, path and everything occur at the same round. Also added unit tests
This commit also include a new Update with 1 extra frame at the end + enabling multiple messages to each player.